- #24 オプション画面の背景途切れを直した - #25 サインアップ情報をSAVEした際にダイアログが出るようにした。 - #26 CREATEボタンを押したとき、サインアップ情報が未登録でもサインアップ画面に遷移できるようにした。 - オプション画面のコンポーネント分け（リファクタリング）
