+++ src/options/App.vue
@@ -14,8 +14,6 @@ body{
   width: 600px !important;
   min-width: 600px !important;
   max-width: 600px !important;
-  height: 500px !important;
-  max-height: 500px !important;
   min-height: 500px !important;
   background-color: lightgrey;
 }
